backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
openai==0.28.1
graphviz==0.20.1